refactor(config-ui): the create blueprint page
