+++ config-ui/src/pages/blueprint/create/api.ts
@@ -18,18 +18,13 @@
 
 import request from '@/components/utils/request'
 
-export const updateDataScope = (
-  plugin: string,
-  connectionId: ID,
-  payload: any
-) =>
-  request(`/plugins/${plugin}/connections/${connectionId}/scopes`, {
-    method: 'put',
-    data: payload
-  })
-
 export const createBlueprint = (payload: any) =>
   request('/blueprints', {
     method: 'post',
     data: payload
+  })
+
+export const runBlueprint = (id: ID) =>
+  request(`/blueprints/${id}/trigger`, {
+    method: 'post'
   })
