+++ config-ui/src/pages/blueprint/create/step-one/styled.ts
@@ -19,8 +19,6 @@
 import styled from 'styled-components'
 import { Colors } from '@blueprintjs/core'
 
-export * from '../styled'
-
 export const ConnectionList = styled.ul`
   margin: 0;
   padding: 12px;
