--- config-ui/src/pages/blueprint/create/step-three/use-transformation.ts
@@ -1,64 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-import { useState, useMemo } from 'react'
-
-import type { Plugins } from '@/plugins'
-import { operator } from '@/utils'
-
-import * as API from '../api'
-
-interface Props {
-  plugin: Plugins
-  connectionId: ID
-  onBack: () => void
-}
-
-export const useTransformation = ({ plugin, connectionId, onBack }: Props) => {
-  const [saving, setSaving] = useState(false)
-  const [selectedScope, setSelectedScope] = useState<any>([])
-
-  const updateScopeRule = async (tsId: ID) => {
-    setSaving(true)
-    const paylod = selectedScope.map((sc: any) => ({
-      ...sc,
-      transformationRuleId: tsId
-    }))
-
-    if (paylod.length) {
-      await operator(() =>
-        API.updateDataScope(plugin, connectionId, {
-          data: paylod
-        })
-      )
-    }
-
-    setSaving(false)
-    onBack()
-  }
-
-  return useMemo(
-    () => ({
-      saving,
-      selectedScope,
-      onChangeSelectedScope: setSelectedScope,
-      updateScopeRule
-    }),
-    [saving, selectedScope]
-  )
-}
